requirements.txt
Django==2.0.5
pytz==2018.5
pyyaml==3.13